% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createGenind}
\alias{createGenind}
\title{Convert input file to genind object}
\usage{
createGenind(Ifile, ploidy)
}
\arguments{
\item{Ifile}{Input file object.}

\item{ploidy}{Ploidy ("Haploid" or "Diploid").}
}
\value{
An object of class genind.
}
\description{
Convert input file to genind object
}
\keyword{internal}
