% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{data_Server}
\alias{data_Server}
\title{Generate input data tab Server.}
\usage{
data_Server(
  id,
  getgenind,
  getData,
  barplotcolor,
  transparency,
  width,
  height,
  popnames
)
}
\description{
Generate input data tab Server.
}
\keyword{internal}
