% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_conversion.R
\name{file_conv_Server}
\alias{file_conv_Server}
\title{Generate file conversion tab Server}
\usage{
file_conv_Server(id, fpath, ploidy, getgenind, popnames)
}
\description{
Generate file conversion tab Server
}
\keyword{internal}
