% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_for_popgen.R
\name{for_popgen_Server}
\alias{for_popgen_Server}
\title{Generate the forensic parameters and population genetics tabs Server.}
\usage{
for_popgen_Server(
  id,
  input_file,
  getgenind,
  popnames,
  ploidy,
  hw_perm,
  barplotcolor,
  transparency,
  cexaxis
)
}
\description{
Generate the forensic parameters and population genetics tabs Server.
}
\keyword{internal}
