% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_reference_mds.R
\name{freq_to_mds}
\alias{freq_to_mds}
\title{Convert allele frequencies to a format suitable for MDS analysis.}
\usage{
freq_to_mds(fname)
}
\description{
Convert allele frequencies to a format suitable for MDS analysis.
}
\keyword{internal}
