% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_for_popgen.R
\name{get_2p2q2}
\alias{get_2p2q2}
\title{2p2q2}
\usage{
get_2p2q2(a, b)
}
\arguments{
\item{a}{first value}

\item{b}{second value}
}
\value{
Computes 2 * a^2 * b^2
}
\description{
2p2q2
}
\keyword{internal}
