% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_hw_gp}
\alias{get_hw_gp}
\title{Parse HWE test Genepop results.}
\usage{
get_hw_gp(fname, n_iter)
}
\arguments{
\item{fname}{path to genepop file}
}
\value{
a dataframe containing HWE test p-values
}
\description{
Parse HWE test Genepop results.
}
\keyword{internal}
