% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ld_gp}
\alias{get_ld_gp}
\title{Parse Linkage Disequilibrium Genepop results.}
\usage{
get_ld_gp(out.name)
}
\arguments{
\item{out.name}{path to genepop file}
}
\value{
a dataframe containing LD test p-values
}
\description{
Parse Linkage Disequilibrium Genepop results.
}
\keyword{internal}
