% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFreqAllPop}
\alias{getFreqAllPop}
\title{Get allele frequencies for each population.}
\usage{
getFreqAllPop(data)
}
\arguments{
\item{data}{a genind object}
}
\value{
an allele frequency table
}
\description{
Get allele frequencies for each population.
}
\keyword{internal}
