% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getFreqFromGenind}
\alias{getFreqFromGenind}
\title{Get allele frequencies for a given population.}
\usage{
getFreqFromGenind(data)
}
\arguments{
\item{data}{a genind object}
}
\value{
an allele frequency table
}
\description{
Get allele frequencies for a given population.
}
\keyword{internal}
