% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGenepopHW}
\alias{getGenepopHW}
\title{Hardy-Weinberg Equilibrium test..}
\usage{
getGenepopHW(data, ploidy, n_iter = 10000)
}
\arguments{
\item{data}{path to straf input file}

\item{ploidy}{ploidy}
}
\value{
a table containing Genepop's HW test p-values
}
\description{
Hardy-Weinberg Equilibrium test..
}
\keyword{internal}
