% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGenepopLD}
\alias{getGenepopLD}
\title{Pairwise Linkage Disequilibrium computation.}
\usage{
getGenepopLD(data, ploidy, n_iter = 10000)
}
\arguments{
\item{data}{a genind object}
}
\value{
a table containing Genepop's LD p-values
}
\description{
Pairwise Linkage Disequilibrium computation.
}
\keyword{internal}
