% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_haplo_stats.R
\name{getHaploStatsAllPop}
\alias{getHaploStatsAllPop}
\title{Get haplotype specific statistics for all populations}
\usage{
getHaploStatsAllPop(data)
}
\arguments{
\item{data}{A genind object.}
}
\description{
Get haplotype specific statistics for all populations
}
\keyword{internal}
