% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_haplo_stats.R
\name{getHaploStatsFromGenind}
\alias{getHaploStatsFromGenind}
\title{Get haplotype specific statistics}
\usage{
getHaploStatsFromGenind(data)
}
\arguments{
\item{data}{A genind object.}
}
\description{
Get haplotype specific statistics
}
\keyword{internal}
