% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_for_popgen.R
\name{getIndicesFromGenind}
\alias{getIndicesFromGenind}
\title{Get forensics and popgen indices for a given population}
\usage{
getIndicesFromGenind(data, ploidy = 2)
}
\arguments{
\item{data}{A genind object.}

\item{ploidy}{character, ploidy, either "Diploid" or "Haploid".}
}
\description{
Get forensics and popgen indices for a given population
}
\keyword{internal}
