% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_hap_dist}
\alias{plot_hap_dist}
\title{Plot pairwise distance between haplotypes.}
\usage{
plot_hap_dist(df)
}
\arguments{
\item{df}{a table containing pairwise distance between haplotypes}
}
\value{
a heatmap representing pairwise distance between haplotypes
}
\description{
Plot pairwise distance between haplotypes.
}
\keyword{internal}
