% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_hap_div}
\alias{plot_hap_div}
\title{Plot Pairwise Distance between haplotypes.}
\usage{
plot_hap_div(df)
}
\arguments{
\item{df}{a table containing LD test results}
}
\value{
a heatmap representing LD test p-values
}
\description{
Plot Pairwise Distance between haplotypes.
}
\keyword{internal}
