% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_ld}
\alias{plot_ld}
\title{Plot Pairwise Linkage Disequilibrium.}
\usage{
plot_ld(df)
}
\arguments{
\item{df}{a table containing LD test results}

\item{pop}{string, population to represent}
}
\value{
a heatmap representing LD test p-values
}
\description{
Plot Pairwise Linkage Disequilibrium.
}
\keyword{internal}
