% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/module_file_conversion.R
\name{poptree2straf}
\alias{poptree2straf}
\title{Convert a POPTREE file to a format suitable for use in STRAF}
\usage{
poptree2straf(input_file, output_file)

poptree2straf(input_file, output_file)
}
\arguments{
\item{fname}{Input file name in the STRAF format.}
}
\value{
An object of class genind.
}
\description{
Convert a POPTREE file to a format suitable for use in STRAF

Convert POPTREE file to the STRAF allele frequencies format.
This function
}
\keyword{internal}
