% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_reference_mds.R
\name{ref_mds_Server}
\alias{ref_mds_Server}
\title{Generate the reference population tab Server.}
\usage{
ref_mds_Server(id, getgenind)
}
\description{
Generate the reference population tab Server.
}
\keyword{internal}
