% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStraf.R
\name{runStraf}
\alias{runStraf}
\title{Run the STRAF application}
\usage{
runStraf()
}
\value{
Runs the shiny application.
}
\description{
Main function to be called in order to start STRAF,
}
\examples{
# runStraf()

}
