% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straf.R
\docType{package}
\name{straf}
\alias{straf}
\title{straf: STR Analysis for Forensics}
\description{
straf is a Shiny application to perform Short Tandem Repeats (STRs, also
known as microsatellites) data analysis. The application allows one to
compute forensic parameters, population genetics indices, and investigate
population structure through various methods and generate relevant data
visualisations. It also implements file conversion to other popular formats.
}
\section{Running the app}{

One simply needs to call the runStraf() function to start the application.
}

