% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_conversion.R
\name{straf2arlequin}
\alias{straf2arlequin}
\title{Convert STRAF file to the Arlequin format.
This function}
\usage{
straf2arlequin(input_file, output_file)
}
\arguments{
\item{fname}{Input file name in the STRAF format.}
}
\description{
Convert STRAF file to the Arlequin format.
This function
}
\keyword{internal}
