% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_conversion.R
\name{straf2familias}
\alias{straf2familias}
\title{Convert STRAF file to the Familias format.}
\usage{
straf2familias(input_file, output_file)
}
\description{
Convert STRAF file to the Familias format.
}
\keyword{internal}
