% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_file_conversion.R
\name{straf2genepop}
\alias{straf2genepop}
\title{Convert STRAF file to the Genepop format.}
\usage{
straf2genepop(input_file, output_file, ploidy = 2)
}
\description{
Convert STRAF file to the Genepop format.
}
\keyword{internal}
