% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStraf.R
\name{runStraf}
\alias{runStraf}
\title{Run the STRAF application}
\usage{
runStraf(port = 3838, host = "0.0.0.0")
}
\arguments{
\item{port}{TCP port.}

\item{host}{IP address.}
}
\value{
Runs the shiny application.
}
\description{
Main function to be called in order to start STRAF,
}
\examples{
library(straf)
# runStraf()

}
